# Faster string generator