--- a/.gitignore
+++ b/.gitignore
@@ -8,7 +8,8 @@
 /spec/examples.txt
 /test/tmp/
 /test/version_tmp/
-/tmp/
+/tmp/*
+!/tmp/.keep
 
 # Used by dotenv library to load environment variables.
 # .env