--- a/benchmark/compare_with_othergems/kachick/Gemfile
+++ b/benchmark/compare_with_othergems/kachick/Gemfile
@@ -2,5 +2,5 @@
 
 source('https://rubygems.org')
 
-gem('ruby-ulid', '0.5.0')
+gem('ruby-ulid', path: '../../../')
 gem('benchmark-ips', '~> 2.10.0')