--- a/benchmark/generate_vs_gen.rb
+++ b/benchmark/generate_vs_gen.rb
@@ -1,0 +1,11 @@
+# coding: utf-8
+# frozen_string_literal: true
+
+require('benchmark/ips')
+require_relative('../lib/ulid')
+
+Benchmark.ips do |x|
+  x.report('ULID.generate.to_s') { ULID.generate.to_s }
+  x.report('ULID.gen') { ULID.gen }
+  x.compare!
+end