--- a/lib/ulid/monotonic_generator.rb
+++ b/lib/ulid/monotonic_generator.rb
@@ -70,6 +70,10 @@ class ULID
       end
     end
 
+    # @TODO Consider to provide this
+    # def encode
+    # end
+
     undef_method(:freeze)
 
     # @raise [TypeError] always raises exception and does not freeze self