--- a/lib/ulid/version.rb
+++ b/lib/ulid/version.rb
@@ -3,5 +3,5 @@
 # shareable_constant_value: literal
 
 class ULID
-  VERSION = '0.5.0'
+  VERSION = '0.6.0.pre'
 end