--- a/scripts/prof.rb
+++ b/scripts/prof.rb
@@ -4,6 +4,10 @@
 require 'stackprof'
 require_relative '../lib/ulid'
 
+StackProf.run(mode: :wall, out: "./tmp/stackprof-wall-#{Time.now.to_i}.dump") do
+  100000.times { ULID.encode }
+end
+
 StackProf.run(mode: :cpu, out: "./tmp/stackprof-cpu-#{Time.now.to_i}.dump") do
   100000.times { ULID.encode }
 end