--- a/test/core/test_ulid_class.rb
+++ b/test/core/test_ulid_class.rb
@@ -30,6 +30,7 @@ class TestULIDClass < Test::Unit::TestCase
         :min,
         :milliseconds_from_moment,
         :generate,
+        :gen,
         :from_integer,
         :normalize,
         :floor,
@@ -597,6 +598,34 @@ class TestULIDClass < Test::Unit::TestCase
     [nil, 4.2, 42/24r, '42', ulid, ulid.to_s, BasicObject.new, Object.new].each do |evil|
       err = assert_raises(ArgumentError) do
         ULID.generate(moment: evil)
+      end
+      assert_equal('`moment` should be a `Time` or `Integer as milliseconds`', err.message)
+    end
+  end
+
+  def test_gen
+    assert_instance_of(String, ULID.gen)
+    assert_not_equal(ULID.gen, ULID.gen)
+
+    time = Time.at(946684800, Rational('123456.789')).utc
+    gen = ULID.gen(moment: time)
+    ulid = ULID.parse(gen)
+    assert_equal(ulid.to_s, gen)
+    assert_not_same(ulid.to_s, gen)
+    assert_equal(gen.frozen?, ulid.to_s.frozen?)
+    assert_true(gen.frozen?)
+    assert_not_equal(time, ulid.to_time)
+    assert_equal(true, ulid.to_time < time)
+    assert_equal(ULID.floor(time), ulid.to_time)
+    milliseconds = 42
+    assert_equal(Time.at(0, milliseconds, :millisecond), ULID.parse(ULID.gen(moment: milliseconds)).to_time)
+
+    entropy = 42
+    assert_equal(entropy, ULID.parse(ULID.gen(entropy: entropy)).entropy)
+
+    [nil, 4.2, 42/24r, '42', ulid, ulid.to_s, BasicObject.new, Object.new].each do |evil|
+      err = assert_raises(ArgumentError) do
+        ULID.gen(moment: evil)
       end
       assert_equal('`moment` should be a `Time` or `Integer as milliseconds`', err.message)
     end